region on c

+++ contiki-ng/arch/platform/cooja/dev/position_intf.h
@@ -34,6 +34,6 @@
 extern double X;
 extern double Y;
 extern double Z;
-extern double REGION;
+extern int REGION;
 
 #endif /* POSITION_INTF_H_ */
